Merged file and image message types
Until this commit, messages with images were very similar to messages
with files, as they both required uploading to the same file endpoint.
But after uploading, the image was added to the message as markdown,
whilst the file message used a different file type. This is a lot more
flexible, since it offers an array with some metadata.

This discrepancy also forced us to maintain two very similar functions
to upload both images and files, which gave no benefit. For both these
reasons, we decided to remove the image messages and let the clients
figure out how they want to display file images, based on the `type`
metadata. Future commits will improve this UI, but this commit simply
recreates the old image UI that was used until now, but with the new
file upload system.